Update eslint and eslint-plugin-react. (#136)
* Update eslint and eslint-plugin-react.

* stylelint: add 'include' to 'ignoreAtRules' for 'at-rule-empty-line-before'

This is because the stylelinter fails after adding _type.scss. For some reason, it seems like stylelinter believes that @import and @include are the same type of at-rule, an therefore doesn't want a blank line between them.

Admittedly, this change is not related to the upgrade of eslint from version 2 to 3, but include it here anyway, since it makes the build fail.