Change fqdn to beta.oppgaver.kidsakoder.no (CNAME). Also, change all keywords from dash to underscore, as shown in the documentation. (#666)
It seems like dash also works, but that might change in the future. In any case, might as well have the same as they have in the docs.
This commits also moves git config diff.renameLimit to before_deploy, and comments it out. If it doesn't work without it, uncomment it.